Move to double-buffer based rendering
Our default mode is to soft-wrap lines exceeding the viewport width.
This means that just by looking at buffer lines is not enough to make
a good decision on how many buffer lines we need to jump so scroll,
say, "half a screen".

However, if we render the current viewport with some margin before
and after, then we can more easily calculate how many buffer lines
we need to jump to make a visual jump.