highlighting & ui/editor: Better interface & handle keywords better
The changes in 'highlighting' are mainly to support proper keyword
matching and arbitrary regexp patterns. 'FindAllStringSubmatchIndex'
was in fact handling only non-overlapping matches (as documented)
and this was not the case we were looking for.

The (temporarily) builtin patterns in 'editor' do not handle
comments within quotes correctly. To be continued.